Add Google Cloud SQL support for App Engine.
It uses Cloud SQL as a database if it runs on Google App Engine with
using cloudsql DSN.

appengine.go includes a build constraint for Google App Engine (GAE)
(http://blog.golang.org/the-app-engine-sdk-and-workspaces-gopath) to
avoid an error outside App Engine and provide an exception. Using
Google Cloud SQL with App Engine GO SDK is not supported yet. It is a
SDK limitation.