Switch to rolling release
The `go1` tag caused a lot of confusion and did more harm than good.
In order to make `go get` work as expected and reduce the amount of duplicate bug reports, we switch back to a "rolling release". `go get` will install the current master branch again.
But we will continue to create branches for stable versions.