Use hexadecimal literals for []byte data.
Since 5.6.27, mysql server produces a warning if string literals cannot
be interpreted in a given character set (see Bug #20238729).
As a consequence, the driver may fail to insert []byte data into
BLOB columns when interpolateParams is true.

This commit changes mysqlConn.interpolateParams so that []byte data
are interpolated as hexadecimal literals instead of string literals.

https://dev.mysql.com/doc/refman/5.7/en/hexadecimal-literals.html