Change ErrUnknownPlugin to a func which can take the plugin name
MySQL 8.0 brings a new authentication plugin which is enabled by
default. This plugin is not yet handled by the go driver.

The error message which is returned is somewhat vague and this is
likely to cause confusion.

The change provides the same error message but includes the plugin
name which will be easier for people to lookup. So rather than
returning the error:

this authentication plugin is not supported

we return something like:

authentication plugin "caching_sha2_password" is not supported