move tls config logic to normalize
With the driver.Connector interface you can skip using string based DSNs
and use a Config object directly with Connector.  However the logic to
create the tls.Config is only found in parseDSNParams.

This keeps the string parsing logic in parseDSNParams but moves the tls
generation logic into Config.normalize.

This is still less than ideal since we cannot directly pass in
tls.Config into Config and have it be used, but it is sill backwards
compatable.  In the future this should be revisited to be able to use a
custome tls.Config passed directly in without string
parsing/registering.