move stale connection check to ResetSession() (#1496)
When ResetSession was added, it was called when the connection is put into the pool.
Thet is why we had only set `mc.reset` flag on ResetSession().

In Go 1.15, this behavior was changed. (golang/go@971f8a2)

ResetSession is called when the connection is checked out from the pool.
So we can call checkConnLiveness() directly from ResetSession.